+++ packages/antd-management-fast-design-playground/CHANGELOG.md
@@ -1,5 +1,16 @@
 # antd-management-fast-design-playground
 
+## 1.0.235
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-design-react@1.0.156
+  - antd-management-fast-component@2.7.79
+  - antd-management-fast-formily@1.0.174
+  - antd-management-fast-common@2.6.184
+
 ## 1.0.234
 
 ### Patch Changes